Improve integer kind handling
Now, all enum define a 'parameter' integer, allowing for
    integer(kind(CHFL_LOG_LEVEL)) => integer(CHFL_LOG_LEVEL)
transformation

Also remove the -fdefault-integer-8 flag to gfortran and explicitlty
specify the size of all the integers. c_int is assumed to be the same
size as int32 here, there should be some very strange platforms on which
this is not true.